+++ packages/plugin-packages/mashroom-security-provider-openid-connect/test/create-user.test.ts
@@ -13,7 +13,7 @@ describe('create-user', () => {
             sub: 'admin'
         };
 
-        const user = createUser(claims, null, null);
+        const user = createUser(claims, null, null, undefined, null);
 
         expect(user).toEqual({
             username: 'admin',
@@ -38,7 +38,7 @@ describe('create-user', () => {
             email: 'admin@test.com',
         };
 
-        const user = createUser(claims, null, null);
+        const user = createUser(claims, null, null, undefined, null);
 
         expect(user).toEqual({
             username: 'admin',
@@ -64,7 +64,7 @@ describe('create-user', () => {
             roles: ['Role1', 'Role2', 'Role3'],
         };
 
-        const user = createUser(claims, null, 'roles', ['Role2']);
+        const user = createUser(claims, null, 'roles', ['Role2'], null);
 
         expect(user).toEqual({
             username: 'admin',
@@ -94,7 +94,7 @@ describe('create-user', () => {
             roles: ['Role1', 'Role2', 'Role3'],
         };
 
-        const user = createUser(claims, userInfo, 'roles', ['Role2']);
+        const user = createUser(claims, userInfo, 'roles', ['Role2'], null);
 
         expect(user).toEqual({
             username: 'admin',
@@ -107,5 +107,73 @@ describe('create-user', () => {
         });
     });
 
+    it('copies extra data from the id token', () => {
+        const claims: IdTokenClaims = {
+            aud: 'x',
+            exp: 0,
+            iat: 0,
+            iss: 'x',
+            sub: 'x',
+            preferred_username: 'admin',
+            name: 'Admin User',
+            email: 'admin@test.com',
+            roles: ['Role1', 'Role2', 'Role3'],
+            phone: '123123123',
+        };
 
+        const user = createUser(claims, undefined, 'roles', ['Role2'], {
+            phoneNumber: 'phone',
+            foo: 'foo',
+        });
+
+        expect(user).toEqual({
+            username: 'admin',
+            displayName: 'Admin User',
+            email: 'admin@test.com',
+            pictureUrl: null,
+            extraData: {
+                foo: null,
+                phoneNumber: '123123123',
+            },
+            roles: ['Role1', 'Role2', 'Role3', 'Administrator'],
+            secrets: null,
+        });
+    });
+
+    it('copies extra data from the user info', () => {
+        const claims: IdTokenClaims = {
+            aud: 'x',
+            exp: 0,
+            iat: 0,
+            iss: 'x',
+            sub: 'admin',
+        };
+        const userInfo: UserinfoResponse = {
+            sub: 'admin',
+            preferred_username: 'admin',
+            name: 'Admin User',
+            email: 'admin@test.com',
+            picture: 'http://my-picture.com',
+            roles: ['Role1', 'Role2', 'Role3'],
+            phone: '222333',
+        };
+
+        const user = createUser(claims, userInfo, 'roles', ['Role2'], {
+            phoneNumber: 'phone',
+            foo: 'foo',
+        });
+
+        expect(user).toEqual({
+            username: 'admin',
+            displayName: 'Admin User',
+            email: 'admin@test.com',
+            pictureUrl: 'http://my-picture.com',
+            extraData: {
+                foo: null,
+                phoneNumber: '222333',
+            },
+            roles: ['Role1', 'Role2', 'Role3', 'Administrator'],
+            secrets: null,
+        });
+    });
 });